pyproject.toml
[tool.poetry]
name = "scrapus"
version = "0.1.0"
description = "A scraper for Dofus.com"
authors = ["cjw21824"]

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.6.5"
SQLAlchemy = "^1.4.21"
selenium = "^3.141.0"
beautifulsoup4 = "^4.9.3"
lxml = "^4.6.3"
azure-core = "^1.16.0"
azure-storage-blob = "^12.8.1"
azure-keyvault-secrets = "^4.3.0"
azure-identity = "^1.6.0"
python-dotenv = "^0.19.0"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
